requirements.txt
numpy
scipy
scikit-image
scikit-learn
pytorch
torchvision
tqdm
click
pillow
opencv
